TODO file added and minor change to hostname checking error message
A TODO file was added to track needed bugfixes and pending feature requests.  A minor update to the error message displayed when performing hostname checking was also performed to clarify the correct way to configure a prx file when hostname checking is not desired.  This is important because the message displayed when the prx:hostname tag is present but does not contain a value is confusing.